+++ ehpcinstant-20230701/src/Models/ListJobExecutorsResponseBody/executors.php
@@ -30,6 +30,11 @@ class executors extends Model
     public $endTime;
 
     /**
+     * @var string
+     */
+    public $executorId;
+
+    /**
      * @var string[]
      */
     public $hostName;
@@ -56,6 +61,7 @@ class executors extends Model
         'arrayIndex'   => 'ArrayIndex',
         'createTime'   => 'CreateTime',
         'endTime'      => 'EndTime',
+        'executorId'   => 'ExecutorId',
         'hostName'     => 'HostName',
         'ipAddress'    => 'IpAddress',
         'status'       => 'Status',
@@ -77,6 +83,9 @@ class executors extends Model
         }
         if (null !== $this->endTime) {
             $res['EndTime'] = $this->endTime;
+        }
+        if (null !== $this->executorId) {
+            $res['ExecutorId'] = $this->executorId;
         }
         if (null !== $this->hostName) {
             $res['HostName'] = $this->hostName;
@@ -110,6 +119,9 @@ class executors extends Model
         }
         if (isset($map['EndTime'])) {
             $model->endTime = $map['EndTime'];
+        }
+        if (isset($map['ExecutorId'])) {
+            $model->executorId = $map['ExecutorId'];
         }
         if (isset($map['HostName'])) {
             if (!empty($map['HostName'])) {